+++ CHANGELOG.md
@@ -8,11 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- if expression.
+
 ### Changed
 - `FunctionError` type is replaced with `anyhow::Error`
 
 ### Removed
-- Methods for adding rules and functions to a `RuleSet`, use the builder
+- Methods for adding rules and functions to a `RuleSet` directly.
 
 
 ## [0.4.4] - 2022-12-08