+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [0.4.5] - 2022-12-29
 
 ### Added
 - if expression.
@@ -126,7 +126,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - async `UserFunction` support
 
 
-[unreleased]: https://github.com/mendelt/reval/compare/v0.4.4...HEAD
+[unreleased]: https://github.com/mendelt/reval/compare/v0.4.5...HEAD
+[0.4.5]: https://github.com/mendelt/reval/compare/v0.4.4...v0.4.5
 [0.4.4]: https://github.com/mendelt/reval/compare/v0.4.3...v0.4.4
 [0.4.3]: https://github.com/mendelt/reval/compare/v0.4.2...v0.4.3
 [0.4.2]: https://github.com/mendelt/reval/compare/v0.4.1...v0.4.2